+++ apps/prairielearn/src/models/shared-state-value.sql
@@ -1,3 +1,11 @@
+-- BLOCK select_assessment_instance_id_for_instance_question
+SELECT
+  iq.assessment_instance_id
+FROM
+  instance_questions AS iq
+WHERE
+  iq.id = $instance_question_id;
+
 -- BLOCK select_value
 SELECT
   *
